feat(votos): añade empate neutral (#195)
* feat(votos): añade empate neutral

* fix(votos): soporta empate neutral en similares

* fix(votos): reparte empate en stats

+++ backend/src/main/java/com/diegoalegil/animeshowdown/dto/BracketUpdateEvent.java
@@ -15,8 +15,8 @@ public record BracketUpdateEvent(
         Long torneoId,
         Long enfrentamientoId,
         Long personaje1Id,
-        long personaje1Votos,
+        double personaje1Votos,
         Long personaje2Id,
-        long personaje2Votos,
-        long totalVotos) {
+        double personaje2Votos,
+        double totalVotos) {
 }
