+++ backend/src/main/java/com/diegoalegil/animeshowdown/dto/TopPersonajeItem.java
@@ -9,5 +9,5 @@ public record TopPersonajeItem(
         String nombre,
         String imagenUrl,
         String anime,
-        long votos) {
+        double votos) {
 }
